refactor: 불필요한 RequestAPI 프로토콜 삭제
* ReuqestAPI 파일에 있던 HTTPMethod enum을 Request 파일로 이동
